+++ flink-scala/src/test/scala/org/apache/flink/api/scala/io/CollectionInputFormatTest.scala
