+++ flink-scala/src/test/scala/org/apache/flink/api/scala/runtime/EnumValueComparatorTest.scala
