+++ flink-tests/src/test/scala/org/apache/flink/api/scala/runtime/CaseClassNormalizedKeySortingTest.scala
@@ -18,9 +18,6 @@
 
 package org.apache.flink.api.scala.runtime
 
-import java.util
-import java.util.{ArrayList, List, Random}
-
 import org.apache.flink.api.common.ExecutionConfig
 import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.apache.flink.api.common.typeutils.{CompositeType, TypeComparator}
@@ -28,80 +25,75 @@ import org.apache.flink.api.scala._
 import org.apache.flink.core.memory._
 import org.apache.flink.runtime.operators.sort.{NormalizedKeySorter, QuickSort}
 
+import java.util
+import java.util.Random
+
 import org.junit.Assert._
 import org.junit.Test
 import org.mockito.Mockito
 
-class CaseClassComparatorTest {
+/** Test that verifies that the case class comparators properly
+ * generate normalized keys used during sorting.
+ */
+class CaseClassNormalizedKeySortingTest {
 
   case class CaseTestClass(a: Int, b: Int, c: Int, d: String)
-  
+
   @Test
   def testNormalizedKeyGeneration(): Unit = {
-    try {
-      
-      val typeInfo = implicitly[TypeInformation[CaseTestClass]]
-                                     .asInstanceOf[CompositeType[CaseTestClass]]
-      
-      val serializer = typeInfo.createSerializer(new ExecutionConfig)
-      val comparator = new FailingCompareDeserializedWrapper(
-          typeInfo.createComparator(
-            Array[Int](0, 2),
-            Array[Boolean](true, true),
-            0,
-            new ExecutionConfig))
-      
-      assertTrue(comparator.supportsNormalizedKey())
-      assertEquals(8, comparator.getNormalizeKeyLen())
-      assertFalse(comparator.isNormalizedKeyPrefixOnly(8))
-      
-      // validate the failing mock
-      {
-        val in1 : DataInputView = Mockito.mock(classOf[DataInputView])
-        val in2 : DataInputView = Mockito.mock(classOf[DataInputView])
-        
-        try {
-          comparator.compareSerialized(in1, in2)
-          fail("should throw an exception")
-        }
-        catch {
-          case e: UnsupportedOperationException => // fine
-          case ee: Exception => fail("unexpected exception")
-        }
+    val typeInfo = implicitly[TypeInformation[CaseTestClass]]
+      .asInstanceOf[CompositeType[CaseTestClass]]
+
+    val serializer = typeInfo.createSerializer(new ExecutionConfig)
+    val comparator = new FailingCompareDeserializedWrapper(
+      typeInfo.createComparator(
+        Array[Int](0, 2),
+        Array[Boolean](true, true),
+        0,
+        new ExecutionConfig))
+
+    assertTrue(comparator.supportsNormalizedKey())
+    assertEquals(8, comparator.getNormalizeKeyLen())
+    assertFalse(comparator.isNormalizedKeyPrefixOnly(8))
+
+    // validate the failing mock
+    {
+      val in1 : DataInputView = Mockito.mock(classOf[DataInputView])
+      val in2 : DataInputView = Mockito.mock(classOf[DataInputView])
+
+      try {
+        comparator.compareSerialized(in1, in2)
+        fail("should throw an exception")
       }
-      
-      
-      val numMemSegs = 20
-      val memory : util.List[MemorySegment] = new util.ArrayList[MemorySegment](numMemSegs)
-      for (i <- 1 to numMemSegs) {
-        memory.add(MemorySegmentFactory.allocateUnpooledSegment(32*1024))
-      }
-      
-      val sorter : NormalizedKeySorter[CaseTestClass] = new NormalizedKeySorter[CaseTestClass](
-               serializer, comparator, memory)
-      
-      val rnd = new Random()
-      var moreToGo = true
-      var num = 0
-      
-      while (moreToGo) {
-        val next = CaseTestClass(rnd.nextInt(), rnd.nextInt(), rnd.nextInt(), "")
-        moreToGo = sorter.write(next)
-        num += 1
-      }
-      
-      print(num)
-      
-      new QuickSort().sort(sorter)
-    }
-    catch {
-      case e: Exception => {
-        e.printStackTrace()
-        fail(e.getMessage())
+      catch {
+        case e: UnsupportedOperationException => // fine
+        case ee: Exception => fail("unexpected exception")
       }
     }
+
+
+    val numMemSegs = 20
+    val memory : util.List[MemorySegment] = new util.ArrayList[MemorySegment](numMemSegs)
+    for (i <- 1 to numMemSegs) {
+      memory.add(MemorySegmentFactory.allocateUnpooledSegment(32*1024))
+    }
+
+    val sorter : NormalizedKeySorter[CaseTestClass] = new NormalizedKeySorter[CaseTestClass](
+      serializer, comparator, memory)
+
+    val rnd = new Random()
+    var moreToGo = true
+    var num = 0
+
+    while (moreToGo) {
+      val next = CaseTestClass(rnd.nextInt(), rnd.nextInt(), rnd.nextInt(), "")
+      moreToGo = sorter.write(next)
+      num += 1
+    }
+
+    new QuickSort().sort(sorter)
   }
-  
+
   class FailingCompareDeserializedWrapper[T](wrapped: TypeComparator[T]) extends TypeComparator[T] {
 
     def hash(record: T) : Int = wrapped.hash(record)
@@ -109,47 +101,47 @@ class CaseClassComparatorTest {
     def setReference(toCompare: T) = wrapped.setReference(toCompare)
 
     def equalToReference(candidate: T): Boolean = wrapped.equalToReference(candidate)
-    
-    def compareToReference(referencedComparator: TypeComparator[T]): Int 
-      = wrapped.compareToReference(referencedComparator)
-    
-    override def supportsCompareAgainstReference(): Boolean 
-      = wrapped.supportsCompareAgainstReference()
-    
+
+    def compareToReference(referencedComparator: TypeComparator[T]): Int
+    = wrapped.compareToReference(referencedComparator)
+
+    override def supportsCompareAgainstReference(): Boolean
+    = wrapped.supportsCompareAgainstReference()
+
     def compare(first: T, second: T): Int = wrapped.compare(first, second)
-    
+
     def compareSerialized(firstSource: DataInputView, secondSource: DataInputView): Int = {
       throw new UnsupportedOperationException("Not Supported")
     }
-    
+
     def supportsNormalizedKey(): Boolean = wrapped.supportsNormalizedKey()
-    
+
     def supportsSerializationWithKeyNormalization(): Boolean
-      = wrapped.supportsSerializationWithKeyNormalization()
-    
+    = wrapped.supportsSerializationWithKeyNormalization()
+
     def getNormalizeKeyLen(): Int = wrapped.getNormalizeKeyLen()
-    
+
     def isNormalizedKeyPrefixOnly(keyBytes: Int): Boolean
-      = wrapped.isNormalizedKeyPrefixOnly(keyBytes)
-    
+    = wrapped.isNormalizedKeyPrefixOnly(keyBytes)
+
     def putNormalizedKey(record: T, target: MemorySegment, offset: Int, numBytes: Int): Unit
-      = wrapped.putNormalizedKey(record, target, offset, numBytes)
-    
+    = wrapped.putNormalizedKey(record, target, offset, numBytes)
+
     def writeWithKeyNormalization(record: T, target: DataOutputView): Unit
-      = wrapped.writeWithKeyNormalization(record, target)
-    
+    = wrapped.writeWithKeyNormalization(record, target)
+
     def readWithKeyDenormalization(reuse: T, source: DataInputView): T
-      = wrapped.readWithKeyDenormalization(reuse, source)
-    
+    = wrapped.readWithKeyDenormalization(reuse, source)
+
     def invertNormalizedKey(): Boolean = wrapped.invertNormalizedKey()
-    
+
     def duplicate(): TypeComparator[T] = new FailingCompareDeserializedWrapper(wrapped.duplicate())
-    
+
     def extractKeys(record: Object, target: Array[Object], index: Int): Int
-      = wrapped.extractKeys(record, target, index)
-    
+    = wrapped.extractKeys(record, target, index)
+
     def getFlatComparators(): Array[TypeComparator[_]] = wrapped.getFlatComparators()
-    
+
     override def compareAgainstReference(keys: Array[Comparable[_]]): Int = {
       throw new UnsupportedOperationException("Workaround hack.")
     }
